Fix bug in sendResp function
The sendResp function in transport.go was missing a parameter 'isObs' which caused an error when trying to write the response message. This commit adds the 'isObs' parameter to the function signature and checks if it is false before writing the message. This fixes the bug and ensures that the response is only written when necessary.

Signed-off-by: SammyOina <sammyoina@gmail.com>